Refactor code block component
