+++ src/components/code-block/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./code-block"
